requirements.txt
appnope==0.1.0
atomicwrites==1.1.5
attrs==18.1.0
backcall==0.1.0
decorator==4.3.0
Django==2.1
djangorestframework==3.8.2
dnspython==1.15.0
falcon==1.4.1
gunicorn==19.9.0
ipython==6.5.0
ipython-genutils==0.2.0
jedi==0.12.1
more-itertools==4.3.0
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
pluggy==0.7.1
prompt-toolkit==1.0.15
ptyprocess==0.6.0
py==1.5.4
Pygments==2.2.0
pytest==3.7.2
pytest-env==0.6.2
python-etcd==0.4.5
python-mimeparse==1.6.0
pytz==2018.5
PyYAML==3.13
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.23
wcwidth==0.1.7
